+++ CHANGELOG.md
@@ -8,8 +8,10 @@
 - Introduce command "pim:completeness:calculate" size argument to manage number of completenesses to calculate
 - Switching tree to see sub-categories products count and allow filtering on it
 - Group types management
-- Import/Export product groups
-- Import/Export associations
+- Import/Export product groups (CSV)
+- Import/Export associations (CSV)
+- Import/Export attributes (CSV)
+- Import/Export attribute options (CSV)
 - Upload and import an archive
 
 ## Improvements
@@ -26,7 +28,7 @@
 
 ## BC breaks
 - Command "pim:product:completeness-calculator" has been replaced into "pim:completeness:calculate"
-
+- Refactor in ImportExport bundle for Readers, Writers and Processors
 
 # 1.0.0-beta-2 - "Hold the Lion, Please" (2013-10-29)